+++ packages/client/src/index.ts
@@ -1,1 +1,2 @@
 export { MessagePickupRepositoryClient } from './MessagePickupRepositoryClient'
+export { ConnectionInfo } from './interfaces'