+++ packages/server/src/websocket/websocket.gateway.ts
@@ -62,13 +62,13 @@ export class WebsocketGateway implements OnModuleInit, OnModuleDestroy {
         port: WS_PORT,
       })
 
-      this.server.event('messageReceive')
+      this.server.event('messagesReceived')
       this.server.createError(500, 'Error initializing WebSocket server', {
         jsonrpc: '2.0',
         result: 'error',
         id: 1,
       })
-      this.logger.error(`WebSocket server listening on port ${WS_PORT}`)
+      this.logger.log(`WebSocket server listening on port ${WS_PORT}`)
     } catch (error) {
       this.logger.error('Error during WebSocket server initialization', error.stack)
       throw error