+++ packages/server/src/websocket/websocket.module.ts
@@ -5,6 +5,10 @@ import { MongooseModule } from '@nestjs/mongoose'
 import { StoreQueuedMessageSchema, StoreQueuedMessage } from './schemas/StoreQueuedMessage'
 import { StoreLiveSessionSchema, StoreLiveSession } from './schemas/StoreLiveSession'
 import { MessagePersister } from './services/MessagePersister'
+import { HttpModule } from '@nestjs/axios'
+import { FcmNotificationSender } from '../providers/FcmNotificationSender'
+import { PushNotificationQueueService } from '../providers/PushNotificationQueueService'
+import { ApnNotificationSender } from '../providers/ApnNotificationSender'
 
 @Module({
   imports: [
@@ -12,7 +16,15 @@ import { MessagePersister } from './services/MessagePersister'
       { name: StoreQueuedMessage.name, schema: StoreQueuedMessageSchema },
       { name: StoreLiveSession.name, schema: StoreLiveSessionSchema },
     ]),
+    HttpModule,
   ],
-  providers: [WebsocketGateway, WebsocketService, MessagePersister],
+  providers: [
+    WebsocketGateway,
+    WebsocketService,
+    MessagePersister,
+    FcmNotificationSender,
+    ApnNotificationSender,
+    PushNotificationQueueService,
+  ],
 })
 export class WebsocketModule {}