+++ packages/server/test/WebsocketTestModule.ts
@@ -4,6 +4,10 @@ import { WebsocketGateway } from '../src/websocket/websocket.gateway'
 import { MongooseModule } from '@nestjs/mongoose'
 import { StoreQueuedMessageSchema, StoreQueuedMessage } from '../src/websocket/schemas/StoreQueuedMessage'
 import { StoreLiveSessionSchema, StoreLiveSession } from '../src/websocket/schemas/StoreLiveSession'
+import { HttpModule } from '@nestjs/axios'
+import { FcmNotificationSender } from '../src/providers/FcmNotificationSender'
+import { ApnNotificationSender } from '../src/providers/ApnNotificationSender'
+import { PushNotificationQueueService } from '../src/providers/PushNotificationQueueService'
 
 @Module({
   imports: [
@@ -11,7 +15,14 @@ import { StoreLiveSessionSchema, StoreLiveSession } from '../src/websocket/schem
       { name: StoreQueuedMessage.name, schema: StoreQueuedMessageSchema },
       { name: StoreLiveSession.name, schema: StoreLiveSessionSchema },
     ]),
+    HttpModule,
   ],
-  providers: [WebsocketGateway, WebsocketService],
+  providers: [
+    WebsocketGateway,
+    WebsocketService,
+    FcmNotificationSender,
+    ApnNotificationSender,
+    PushNotificationQueueService,
+  ],
 })
 export class WebsocketTestModule {}