+++ packages/server/CHANGELOG.md
@@ -1,15 +1,117 @@
 # Changelog
 
-## [0.0.3](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server@v0.0.2...@2060.io/message-pickup-repository-server@v0.0.3) (2024-10-21)
+## [0.0.8](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server@v0.0.7...@2060.io/message-pickup-repository-server@v0.0.8) (2024-11-13)
 
 
 ### Bug Fixes
 
-* client types and service methods into server ([#5](https://github.com/2060-io/message-pickup-repository/issues/5)) ([a3ba864](https://github.com/2060-io/message-pickup-repository/commit/a3ba864c9e1e22d29890722459443ad7d78330b9))
+* HttpService Injection in WebsocketService for Push Notification Sending ([#37](https://github.com/2060-io/message-pickup-repository/issues/37)) ([52992d4](https://github.com/2060-io/message-pickup-repository/commit/52992d4d46e8c133f26b8213fa9ea30553eb6a63))
 
-## [0.0.2](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server-v0.0.1...@2060.io/message-pickup-repository-server@v0.0.2) (2024-10-18)
+## [0.0.7](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server@v0.0.6...@2060.io/message-pickup-repository-server@v0.0.7) (2024-11-08)
+
+
+### Features
+
+* support message retrieval limit based on byte count ([#32](https://github.com/2060-io/message-pickup-repository/issues/32)) ([0dc06f2](https://github.com/2060-io/message-pickup-repository/commit/0dc06f2ba6d4f3379ae05cca2d76f56bf66b9669))
+
+## [0.0.6](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server@v0.0.5...@2060.io/message-pickup-repository-server@v0.0.6) (2024-11-06)
+
+
+### Features
+
+* Add handle instance execution to Message Persister service ([#30](https://github.com/2060-io/message-pickup-repository/issues/30)) ([a806663](https://github.com/2060-io/message-pickup-repository/commit/a806663a0fe9cbd833f7d87d6fff47817e08700e))
+
+## [0.0.5](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server@v0.0.5...@2060.io/message-pickup-repository-server@v0.0.5) (2024-10-31)
+
+
+### Features
+
+* dynamically build connection to redis cluster ([#22](https://github.com/2060-io/message-pickup-repository/issues/22)) ([9bdb31b](https://github.com/2060-io/message-pickup-repository/commit/9bdb31b556feb4301d759b04182debde1c9b807e))
+
+
+### Bug Fixes
+
+* bump versions ([e5a39aa](https://github.com/2060-io/message-pickup-repository/commit/e5a39aae66abb33c313f51c8cdb796f5a914400a))
+
+## [0.0.5](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server@v0.0.5...@2060.io/message-pickup-repository-server@v0.0.5) (2024-10-31)
+
+
+### Features
+
+* dynamically build connection to redis cluster ([#22](https://github.com/2060-io/message-pickup-repository/issues/22)) ([9bdb31b](https://github.com/2060-io/message-pickup-repository/commit/9bdb31b556feb4301d759b04182debde1c9b807e))
+
+
+### Bug Fixes
+
+* bump versions ([e5a39aa](https://github.com/2060-io/message-pickup-repository/commit/e5a39aae66abb33c313f51c8cdb796f5a914400a))
+
+## [0.0.5](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server@v0.0.5...@2060.io/message-pickup-repository-server@v0.0.5) (2024-10-31)
+
+
+### Features
+
+* dynamically build connection to redis cluster ([#22](https://github.com/2060-io/message-pickup-repository/issues/22)) ([9bdb31b](https://github.com/2060-io/message-pickup-repository/commit/9bdb31b556feb4301d759b04182debde1c9b807e))
+
+
+### Bug Fixes
+
+* bump versions ([e5a39aa](https://github.com/2060-io/message-pickup-repository/commit/e5a39aae66abb33c313f51c8cdb796f5a914400a))
+
+## [0.0.5](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server@v0.0.5...@2060.io/message-pickup-repository-server@v0.0.5) (2024-10-31)
+
+
+### Features
+
+* dynamically build connection to redis cluster ([#22](https://github.com/2060-io/message-pickup-repository/issues/22)) ([9bdb31b](https://github.com/2060-io/message-pickup-repository/commit/9bdb31b556feb4301d759b04182debde1c9b807e))
+
+
+### Bug Fixes
+
+* bump versions ([e5a39aa](https://github.com/2060-io/message-pickup-repository/commit/e5a39aae66abb33c313f51c8cdb796f5a914400a))
+
+## [0.0.5](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server-v0.0.6...@2060.io/message-pickup-repository-server@v0.0.5) (2024-10-30)
 
 
 ### Features
 
 * :sparkles: initial setup to Message Pickup Repository gateway ([#2](https://github.com/2060-io/message-pickup-repository/issues/2)) ([bbf5f20](https://github.com/2060-io/message-pickup-repository/commit/bbf5f207ca63f95f1afe83e17523c23b39c6c841))
+
+
+### Bug Fixes
+
+* bump versions ([e5a39aa](https://github.com/2060-io/message-pickup-repository/commit/e5a39aae66abb33c313f51c8cdb796f5a914400a))
+* client types and service methods into server ([#5](https://github.com/2060-io/message-pickup-repository/issues/5)) ([a3ba864](https://github.com/2060-io/message-pickup-repository/commit/a3ba864c9e1e22d29890722459443ad7d78330b9))
+* ensure single Redis message listener and include new handle messages ([#19](https://github.com/2060-io/message-pickup-repository/issues/19)) ([22848d7](https://github.com/2060-io/message-pickup-repository/commit/22848d7df5a5b8c02055a12b6fda5d299b26ad28))
+* multiple subscriptions to messages received during addLiveSession  ([#17](https://github.com/2060-io/message-pickup-repository/issues/17)) ([37270a0](https://github.com/2060-io/message-pickup-repository/commit/37270a0b21ead05e5c2e2285d001bc8d030c26d4))
+
+## [0.0.6](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server@v0.0.5...@2060.io/message-pickup-repository-server@v0.0.5) (2024-10-29)
+
+
+### Bug Fixes
+
+* bump versions ([e5a39aa](https://github.com/2060-io/message-pickup-repository/commit/e5a39aae66abb33c313f51c8cdb796f5a914400a))
+
+## [0.0.5](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server@v0.0.4...@2060.io/message-pickup-repository-server@v0.0.5) (2024-10-25)
+
+
+### Bug Fixes
+
+* ensure single Redis message listener and include new handle messages ([#19](https://github.com/2060-io/message-pickup-repository/issues/19)) ([22848d7](https://github.com/2060-io/message-pickup-repository/commit/22848d7df5a5b8c02055a12b6fda5d299b26ad28))
+
+## [0.0.4](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server@v0.0.3...@2060.io/message-pickup-repository-server@v0.0.4) (2024-10-23)
+
+### Bug Fixes
+
+- multiple subscriptions to messages received during addLiveSession ([#17](https://github.com/2060-io/message-pickup-repository/issues/17)) ([37270a0](https://github.com/2060-io/message-pickup-repository/commit/37270a0b21ead05e5c2e2285d001bc8d030c26d4))
+
+## [0.0.3](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server@v0.0.2...@2060.io/message-pickup-repository-server@v0.0.3) (2024-10-21)
+
+### Bug Fixes
+
+- client types and service methods into server ([#5](https://github.com/2060-io/message-pickup-repository/issues/5)) ([a3ba864](https://github.com/2060-io/message-pickup-repository/commit/a3ba864c9e1e22d29890722459443ad7d78330b9))
+
+## [0.0.2](https://github.com/2060-io/message-pickup-repository/compare/@2060.io/message-pickup-repository-server-v0.0.1...@2060.io/message-pickup-repository-server@v0.0.2) (2024-10-18)
+
+### Features
+
+- :sparkles: initial setup to Message Pickup Repository gateway ([#2](https://github.com/2060-io/message-pickup-repository/issues/2)) ([bbf5f20](https://github.com/2060-io/message-pickup-repository/commit/bbf5f207ca63f95f1afe83e17523c23b39c6c841))